test/test_digit_recognition.py
import unittest
from unittest.mock import patch
import cv2
import numpy as np
from digit_recognition import DigitRecognizer

class TestDigitRecognizer(unittest.TestCase):
    def setUp(self):
        self.recognizer = DigitRecognizer()

    # 画像が適切に前処理され、指定された形状に変換されていることを確認
    def test_process_image_with_valid_image(self):
        image_path = './data/digits0.png'
        processed_image = self.recognizer.process_image(image_path)
        self.assertIsInstance(processed_image, np.ndarray)

    # 数字認識アルゴリズムが適切に機能していることを確認
    def test_recognize_digits_with_valid_image(self):
        image_path = './data/digits0.png'
        processed_image = self.recognizer.process_image(image_path)
        predicted_string = self.recognizer.recognize_digits(processed_image)
        self.assertIsInstance(predicted_string, str)

    # ファイルが見つからない場合に正しくエラーが処理されることを確認
    def test_process_image_with_invalid_image(self):
        with self.assertRaises(FileNotFoundError):
            self.recognizer.process_image('./invalid/path/to/image.png')

    # 文字が8文字であり、かつ2文字目が2、8文字目が0である　合致するとき認識した数字を出すか確認
    def test_recognize_digits_with_valid_image_and_condition_met(self):
        image_path = './data/digits_condition_met.png'
        processed_image = self.recognizer.process_image(image_path)
        predicted_string = self.recognizer.recognize_digits(processed_image)
        self.assertEqual(predicted_string, '62206250')

    # 文字が8文字であり、かつ2文字目が2、8文字目が0である　合致しないときエラーメッセージを出すか確認
    def test_recognize_digits_with_valid_image_and_condition_not_met(self):
        image_path = './data/digits_condition_not_met.png'
        processed_image = self.recognizer.process_image(image_path)
        predicted_string = self.recognizer.recognize_digits(processed_image)
        self.assertEqual(predicted_string, 'Error: Condition not met.')

    # 他のテストケースの実装
    # ...

if __name__ == '__main__':
    unittest.main()
